Modularize gene structure JavaScript
